Fixed logout on navbar
