+++ src/components/movingbeer/Navbar.vue
@@ -36,7 +36,7 @@
                         </q-item>
                         <q-separator />
                         <q-item clickable v-close-popup>
-                            <q-item-section>
+                            <q-item-section @click="logout()">
                                 <span>
                                     <i class="fas fa-sign-out-alt on-left"></i>Cerrar sesion
                                 </span>
@@ -58,3 +58,23 @@
         </q-toolbar>
     </q-header>
 </template>
+<script>
+import firebase from 'firebase/app'
+import 'firebase/auth'
+export default {
+    methods: {
+        async logout() {
+            firebase
+                .auth()
+                .signOut()
+                .then(async () => {
+                    await this.$store.dispatch('UserLogout')
+                    this.$router.push('/login')
+                })
+                .catch(error => {
+                    console.log(error)
+                })
+        },
+    },
+}
+</script>
